fix(portability): make the Windows MSVC build compile and link
First actual Windows compile (VS 2026, MSVC 14.50 + Ninja) on top of the
Phase 9 Winsock shim. Four MSVC-only fixes; GCC tolerated all four:

- dbisam_catalog.cpp: include duckdb/planner/logical_operator.hpp. MSVC
  needs the complete type to instantiate the deleter for
  unique_ptr<LogicalOperator> (returned/taken by BindCreateIndex);
  catalog.hpp only forward-declares it.
- test_phase2.cpp: qualify dbisam::GUID. The Winsock headers pulled in
  transitively via the socket shim define a global ::GUID that is
  ambiguous against dbisam::GUID under `using namespace dbisam`.
- CMakeLists: set BUILD_UNITTESTS/BUILD_BENCHMARKS OFF before FetchContent.
  We only link libduckdb_static; DuckDB's own test_common.cpp hits a
  std::byte vs <rpcndr.h> ::byte collision on the Windows SDK.
- CMakeLists: set CMAKE_MSVC_RUNTIME_LIBRARY to the static runtime (/MT)
  to match DuckDB, which pins /MT; otherwise the link fails LNK2038.

Also widen a printf in test_phase4_live for the 64-bit Windows socket_t
(was %d on a SOCKET).

Result: build/dbisam.duckdb_extension builds clean. ctest 5/7 — the two
failures (phase2 crypto vector, phase4 catalog-attach body) are golden
byte-capture tests pinned to redacted-out real values on the published
tree, not Windows regressions.

+++ test/test_phase4_live.cpp
@@ -45,8 +45,8 @@ int main() {
 
     try {
         auto client = Client::connect_and_login(opts);
-        std::printf("handshake OK: fd=%d compression=%d batch=%u slot_len=%zu\n",
-                    client.transport().fd(),
+        std::printf("handshake OK: fd=%lld compression=%d batch=%u slot_len=%zu\n",
+                    static_cast<long long>(client.transport().fd()),
                     client.compression() ? 1 : 0,
                     client.batch_size(),
                     client.blob_slot_length());
